convert markdown to HTML
